+++ source/include/Ostap/Formula.h
@@ -34,19 +34,23 @@ namespace Ostap
   public:
     // ========================================================================
     /// constructor from name, expression and the tree 
-    Formula ( const std::string& name       , 
-              const std::string& expression ,
-              TTree*             tree       ) ;
+    Formula 
+    ( const std::string& name       , 
+      const std::string& expression ,
+      TTree*             tree       ) ;
     /// constructor from name, expression and the tree 
-    Formula ( const std::string& name       , 
-              const TCut&        expression ,
-              TTree*             tree       ) ;
+    Formula 
+    ( const std::string& name       , 
+      const TCut&        expression ,
+      TTree*             tree       ) ;
     /// constructor from name, expression and the tree 
-    Formula ( const std::string& expression ,
-              TTree*             tree       ) ;
+    Formula 
+    ( const std::string& expression ,
+      TTree*             tree       ) ;
     /// constructor from name, expression and the tree 
-    Formula ( const TCut&        expression ,
-              TTree*             tree       ) ;
+    Formula 
+    ( const TCut&        expression ,
+      TTree*             tree       ) ;
     /// default constructor, needed for serialisationn 
     Formula () = default  ;
     /// virtual destructor